add user permissions check

+++ controllers/cards.js
@@ -1,6 +1,6 @@
-const Card = require('../models/card');
+const mongoose = require('mongoose');
 
-const isObjectIdValid = require('../validators/object-id-validator');
+const Card = require('../models/card');
 
 async function createCard(req, res) {
   const { name, link } = req.body;
@@ -8,7 +8,7 @@ async function createCard(req, res) {
   try {
     let card = await Card.create({ name, link, owner });
     card = await card.populate(['owner', 'likes']).execPopulate();
-    res.send(card);
+    res.status(201).send(card);
   } catch (err) {
     res.status(500).send({ message: err.message });
   }
@@ -18,33 +18,43 @@ function searchResultHandler(res, card) {
   if (!card) {
     res.status(404).send({ message: 'Карточки с таким id не существует' });
   } else {
-    res.send(card);
+    res.status(200).send(card);
   }
 }
 
-module.exports.checkObjectId = (req, res, next) => {
-  if (!isObjectIdValid(req.params.id)) {
+function searchErrorHandler(res, err) {
+  if (err instanceof mongoose.Error.CastError) {
     res.status(400).send({ message: 'Некорректный id карточки' });
   } else {
-    next();
+    res.status(500).send({ message: err.message });
   }
+}
+
+module.exports.createCard = createCard;
+
+module.exports.deleteCardById = (req, res) => {
+  Card.findById(req.params.id)
+    .then((card) => {
+      if (!card) {
+        res.status(404).send({ message: 'Карточки с таким id не существует' });
+      } else if (card.owner.toString() !== req.user._id) {
+        res.status(403).send({ message: 'Недостаточно прав для удаления данной карточки' });
+      } else {
+        Card.findByIdAndRemove(req.params.id)
+          .populate(['owner', 'likes'])
+          .then((cardToDelete) => searchResultHandler(res, cardToDelete))
+          .catch((err) => searchErrorHandler(res, err));
+      }
+    })
+    .catch((err) => searchErrorHandler(res, err));
 };
 
 module.exports.getCards = (req, res) => {
   Card.find({})
     .populate(['owner', 'likes'])
-    .then((cards) => res.send(cards))
+    .then((cards) => res.status(200).send(cards))
     .catch((err) => res.status(500).send({ message: err.message }));
 };
-
-module.exports.deleteCardById = (req, res) => {
-  Card.findByIdAndRemove(req.params.id)
-    .populate(['owner', 'likes'])
-    .then((card) => searchResultHandler(res, card))
-    .catch((err) => res.status(500).send({ message: err.message }));
-};
-
-module.exports.createCard = createCard;
 
 module.exports.likeCard = (req, res) => {
   Card.findByIdAndUpdate(
@@ -54,7 +64,7 @@ module.exports.likeCard = (req, res) => {
   )
     .populate(['owner', 'likes'])
     .then((card) => searchResultHandler(res, card))
-    .catch((err) => res.status(500).send({ message: err.message }));
+    .catch((err) => searchErrorHandler(res, err));
 };
 
 module.exports.dislikeCard = (req, res) => {
@@ -65,5 +75,5 @@ module.exports.dislikeCard = (req, res) => {
   )
     .populate(['owner', 'likes'])
     .then((card) => searchResultHandler(res, card))
-    .catch((err) => res.status(500).send({ message: err.message }));
+    .catch((err) => searchErrorHandler(res, err));
 };
